Add a feature flag provider_transform_utils, and apply it to the hour cycle transform code
