+++ components/datetime/src/pattern/mod.rs
@@ -3,8 +3,8 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 mod error;
-pub mod hour_cycle;
 mod parser;
+pub mod transform_hour_cycle;
 
 use crate::fields::{self, Field, FieldLength, FieldSymbol};
 pub use error::Error;
@@ -284,5 +284,26 @@ impl Serialize for Pattern {
             }
             seq.end()
         }
+    }
+}
+
+/// Used to represent either H11/H12, or H23/H24. Skeletons only store these
+/// hour cycles as H12 or H24.
+#[derive(Debug, PartialEq, Clone, Copy)]
+#[cfg_attr(
+    feature = "provider_serde",
+    derive(serde::Serialize, serde::Deserialize)
+)]
+pub enum CoarseHourCycle {
+    /// Can either be fields::Hour::H11 or fields::Hour::H12
+    H11H12,
+    /// Can either be fields::Hour::H23 or fields::Hour::H24
+    H23H24,
+}
+
+/// Default is required for serialization, arbitrarily pick one.
+impl Default for CoarseHourCycle {
+    fn default() -> Self {
+        CoarseHourCycle::H11H12
     }
 }
